# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KoproGo API",
    "description": "Belgian Property Management SaaS Platform\n\n# Features\n- 🏢 Building & Unit Management\n- 👥 Multi-owner & Multi-role Support\n- 💰 Financial Management (Belgian PCMN)\n- 🗳️ Meeting & Voting System\n- 📄 Document Management\n- 📊 Budget & État Daté Generation\n- 🔔 Notifications & Payment Recovery\n- 🤝 Community Features (SEL, Notices, Skills)\n- 🎮 Gamification & Achievements\n- 🔐 GDPR Compliant\n\n# Authentication\nAll endpoints (except /health and /public/*) require JWT Bearer token.\nGet token via POST /api/v1/auth/login\n\n# Complete API Documentation\n90 of 511 endpoints annotated with utoipa (Swagger UI live spec).\nFull 495-endpoint OpenAPI 3.0.3 spec available at docs/api/openapi.yaml.\n\nProgressive annotation ongoing — see handlers for pattern.",
    "contact": {
      "name": "KoproGo Support",
      "email": "support@koprogo.com"
    },
    "license": {
      "name": "AGPL-3.0-or-later",
      "url": "https://www.gnu.org/licenses/agpl-3.0.en.html"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development"
    },
    {
      "url": "https://api.koprogo.com",
      "description": "Production"
    }
  ],
  "paths": {
...
    "/buildings/{id}": {
      "get": {
        "tags": [
          "Buildings"
        ],
        "summary": "Get a building by ID",
        "operationId": "get_building",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Building UUID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Building found"
          },
          "404": {
            "description": "Building not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "put": {
        "tags": [
          "Buildings"
        ],
        "summary": "Update a building",
        "operationId": "update_building",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Building UUID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBuildingDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Building updated successfully"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden - SuperAdmin only"
          },
          "404": {
            "description": "Building not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Buildings"
        ],
        "summary": "Delete a building",
        "operationId": "delete_building",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Building UUID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Building deleted successfully"
          },
          "403": {
            "description": "Forbidden - SuperAdmin only"
          },
          "404": {
            "description": "Building not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "UpdateBuildingDto": {
        "type": "object",
        "required": [
          "name",
          "address",
          "city",
          "postal_code",
          "country",
          "total_units"
        ],
        "properties": {
          "acp_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Story 1.2 — Réaffectation de l'ACP parente (SuperAdmin uniquement)."
          },
          "address": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "construction_year": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "country": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "postal_code": {
            "type": "string"
          },
          "total_tantiemes": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "total_units": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from /api/v1/auth/login.\n\nExample: `Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...`\n\nTo authenticate:\n1. Click 'Authorize' button above\n2. Enter token (with or without 'Bearer ' prefix)\n3. Click 'Authorize' in dialog\n4. Try endpoints"
      }
    }
  },
...
}